Refactor: extract searchable list
