+++ lib/pages/settings/dashboards/dashboards_page.dart
@@ -1,130 +1,30 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_svg/svg.dart';
-import 'package:intersperse/intersperse.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
+import 'package:lotti/pages/settings/definitions_list_page.dart';
 import 'package:lotti/services/nav_service.dart';
-import 'package:lotti/themes/theme.dart';
-import 'package:lotti/utils/sort.dart';
-import 'package:lotti/widgets/app_bar/title_app_bar.dart';
 import 'package:lotti/widgets/settings/dashboards/dashboard_definition_card.dart';
-import 'package:lotti/widgets/settings/settings_card.dart';
-import 'package:material_floating_search_bar/material_floating_search_bar.dart';
 
-class DashboardSettingsPage extends StatefulWidget {
+class DashboardSettingsPage extends StatelessWidget {
   const DashboardSettingsPage({super.key});
-
-  @override
-  State<DashboardSettingsPage> createState() => _DashboardSettingsPageState();
-}
-
-class _DashboardSettingsPageState extends State<DashboardSettingsPage> {
-  final JournalDb _db = getIt<JournalDb>();
-
-  late final Stream<List<DashboardDefinition>> stream = _db.watchDashboards();
-  String match = '';
-
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  Widget buildFloatingSearchBar() {
-    final isPortrait =
-        MediaQuery.of(context).orientation == Orientation.portrait;
-
-    final portraitWidth = MediaQuery.of(context).size.width * 0.5;
-
-    return Theme(
-      data: ThemeData(
-        brightness: styleConfig().keyboardAppearance,
-      ),
-      child: FloatingSearchBar(
-        clearQueryOnClose: false,
-        automaticallyImplyBackButton: false,
-        hint: AppLocalizations.of(context)!.settingsDashboardsSearchHint,
-        scrollPadding: const EdgeInsets.only(top: 16, bottom: 56),
-        transitionDuration: const Duration(milliseconds: 800),
-        transitionCurve: Curves.easeInOut,
-        backgroundColor: styleConfig().cardColor,
-        margins: const EdgeInsets.only(top: 8),
-        queryStyle: const TextStyle(
-          fontFamily: mainFont,
-          fontSize: 20,
-        ),
-        hintStyle: TextStyle(
-          fontFamily: mainFont,
-          fontSize: 20,
-          color: styleConfig().secondaryTextColor,
-        ),
-        //physics: const BouncingScrollPhysics(),
-        borderRadius: BorderRadius.circular(8),
-        axisAlignment: isPortrait ? 0 : -1,
-        openAxisAlignment: 0,
-        width: isPortrait ? portraitWidth : MediaQuery.of(context).size.width,
-        onQueryChanged: (query) async {
-          setState(() {
-            match = query.toLowerCase();
-          });
-        },
-        actions: [FloatingSearchBarAction.searchToClear(showIfClosed: false)],
-        builder: (context, transition) {
-          return const SizedBox.shrink();
-        },
-      ),
-    );
-  }
 
   @override
   Widget build(BuildContext context) {
     final localizations = AppLocalizations.of(context)!;
-    void createDashboard() => beamToNamed('/settings/dashboards/create');
 
-    return Scaffold(
-      backgroundColor: styleConfig().negspace,
-      appBar: TitleAppBar(title: localizations.settingsDashboardsTitle),
-      floatingActionButton: FloatingActionButton(
-        backgroundColor: styleConfig().primaryColor,
-        onPressed: createDashboard,
-        child: SvgPicture.asset(styleConfig().actionAddIcon, width: 25),
-      ),
-      body: StreamBuilder<List<DashboardDefinition>>(
-        stream: stream,
-        builder: (
-          BuildContext context,
-          AsyncSnapshot<List<DashboardDefinition>> snapshot,
-        ) {
-          final dashboards = filteredSortedDashboards(
-            snapshot.data ?? [],
-            match: match,
-            showAll: true,
-          );
-
-          return Stack(
-            children: [
-              ListView(
-                shrinkWrap: true,
-                padding: const EdgeInsets.symmetric(vertical: 75),
-                children: intersperse(
-                  const SettingsDivider(),
-                  List.generate(
-                    dashboards.length,
-                    (int index) {
-                      return DashboardDefinitionCard(
-                        dashboard: dashboards.elementAt(index),
-                        index: index,
-                      );
-                    },
-                  ),
-                ).toList(),
-              ),
-              buildFloatingSearchBar(),
-            ],
-          );
-        },
-      ),
+    return DefinitionsListPage<DashboardDefinition>(
+      stream: getIt<JournalDb>().watchDashboards(),
+      createFn: () => beamToNamed('/settings/dashboards/create'),
+      title: localizations.settingsDashboardsTitle,
+      getName: (habitDefinition) => habitDefinition.name,
+      definitionCard: (int index, DashboardDefinition item) {
+        return DashboardDefinitionCard(
+          index: index,
+          dashboard: item,
+        );
+      },
     );
   }
 }
