+++ lib/pages/settings/measurables/measurables_page.dart
@@ -1,138 +1,26 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_svg/svg.dart';
-import 'package:intersperse/intersperse.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
+import 'package:lotti/pages/settings/definitions_list_page.dart';
 import 'package:lotti/services/nav_service.dart';
-import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/app_bar/title_app_bar.dart';
 import 'package:lotti/widgets/settings/measurables/measurable_type_card.dart';
-import 'package:lotti/widgets/settings/settings_card.dart';
-import 'package:material_floating_search_bar/material_floating_search_bar.dart';
 
-class MeasurablesPage extends StatefulWidget {
+class MeasurablesPage extends StatelessWidget {
   const MeasurablesPage({super.key});
-
-  @override
-  State<MeasurablesPage> createState() => _MeasurablesPageState();
-}
-
-class _MeasurablesPageState extends State<MeasurablesPage> {
-  final JournalDb _db = getIt<JournalDb>();
-  String match = '';
-
-  late final Stream<List<MeasurableDataType>> stream =
-      _db.watchMeasurableDataTypes();
-
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  Widget buildFloatingSearchBar() {
-    final isPortrait =
-        MediaQuery.of(context).orientation == Orientation.portrait;
-
-    final portraitWidth = MediaQuery.of(context).size.width * 0.5;
-
-    return Theme(
-      data: ThemeData(
-        brightness: styleConfig().keyboardAppearance,
-      ),
-      child: FloatingSearchBar(
-        clearQueryOnClose: false,
-        automaticallyImplyBackButton: false,
-        hint: AppLocalizations.of(context)!.settingsMeasurablesSearchHint,
-        scrollPadding: const EdgeInsets.only(top: 16, bottom: 56),
-        transitionDuration: const Duration(milliseconds: 800),
-        transitionCurve: Curves.easeInOut,
-        backgroundColor: styleConfig().cardColor,
-        margins: const EdgeInsets.only(top: 8),
-        queryStyle: const TextStyle(
-          fontFamily: mainFont,
-          fontSize: 20,
-        ),
-        hintStyle: TextStyle(
-          fontFamily: mainFont,
-          fontSize: 20,
-          color: styleConfig().secondaryTextColor,
-        ),
-        physics: const BouncingScrollPhysics(),
-        borderRadius: BorderRadius.circular(8),
-        axisAlignment: isPortrait ? 0 : -1,
-        openAxisAlignment: 0,
-        width: isPortrait ? portraitWidth : MediaQuery.of(context).size.width,
-        onQueryChanged: (query) async {
-          setState(() {
-            match = query.toLowerCase();
-          });
-        },
-        actions: [
-          FloatingSearchBarAction.searchToClear(
-            showIfClosed: false,
-          ),
-        ],
-        builder: (context, transition) {
-          return const SizedBox.shrink();
-        },
-      ),
-    );
-  }
 
   @override
   Widget build(BuildContext context) {
     final localizations = AppLocalizations.of(context)!;
-    void createMeasurable() => beamToNamed('/settings/measurables/create');
 
-    return StreamBuilder<List<MeasurableDataType>>(
-      stream: stream,
-      builder: (
-        BuildContext context,
-        AsyncSnapshot<List<MeasurableDataType>> snapshot,
-      ) {
-        final items = snapshot.data ?? [];
-        final filtered = items
-            .where(
-              (MeasurableDataType dataType) =>
-                  dataType.displayName.toLowerCase().contains(match),
-            )
-            .toList();
-
-        return Scaffold(
-          appBar: TitleAppBar(title: localizations.settingsMeasurablesTitle),
-          backgroundColor: styleConfig().negspace,
-          floatingActionButton: FloatingActionButton(
-            backgroundColor: styleConfig().primaryColor,
-            onPressed: createMeasurable,
-            child: SvgPicture.asset(styleConfig().actionAddIcon, width: 25),
-          ),
-          body: Stack(
-            children: [
-              ListView(
-                shrinkWrap: true,
-                padding: const EdgeInsets.only(
-                  bottom: 8,
-                  top: 64,
-                ),
-                children: intersperse(
-                  const SettingsDivider(),
-                  List.generate(
-                    filtered.length,
-                    (int index) {
-                      return MeasurableTypeCard(
-                        item: filtered.elementAt(index),
-                        index: index,
-                      );
-                    },
-                  ),
-                ).toList(),
-              ),
-              buildFloatingSearchBar(),
-            ],
-          ),
-        );
+    return DefinitionsListPage<MeasurableDataType>(
+      stream: getIt<JournalDb>().watchMeasurableDataTypes(),
+      createFn: () => beamToNamed('/settings/measurables/create'),
+      title: localizations.settingsMeasurablesTitle,
+      getName: (dataType) => dataType.displayName,
+      definitionCard: (int index, MeasurableDataType dataType) {
+        return MeasurableTypeCard(index: index, item: dataType);
       },
     );
   }
